New Test Code Developed with Computer Vision Kinematic
New Test Code Developed with Computer Vision Kinematic. This code uses the new inverse kinematic equations developed to allow for the CRS to plug into the port using the OpenMV Camera and April Tags. The new kinematics follow a different motion procedure but the heart of the code is the same as the previous files. This code still needs to be integrated with the OpenMV (still need to solve the problem we are having getting the data from the openMV).